third_party/blink/web_tests/external/wpt/svg/pservers/parsing/stop-opacity-computed-expected.txt

This is a testharness.js-based test.
PASS Property stop-opacity value '-1' computes to '0'
PASS Property stop-opacity value '0.5' computes to '0.5'
PASS Property stop-opacity value '3' computes to '1'
FAIL Property stop-opacity value '-100%' computes to '0' assert_equals: expected "0" but got "1"
FAIL Property stop-opacity value '50%' computes to '0.5' assert_equals: expected "0.5" but got "1"
PASS Property stop-opacity value '300%' computes to '1'
Harness: the test ran to completion.
